Update Main.cpp
Added resizable window and toggle to full screen mode and back. When full screen is toggle, small previous window is drawn on the bottom left for a couple of seconds before the new glViewport is used. This may be an issue, or just a inevitability of switching to full screen at run time.